Moved fetching the articles to new file
