+++ routes/article.js
@@ -4,89 +4,8 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const Article = require('../models/ArticleModel.js')
 const Keyword = require('../models/KeywordModel.js')
-const request = require('request-promise')
-const retext = require('retext')
-const retextkeywords = require('retext-keywords')
-const nlcstToString = require('nlcst-to-string')
 
 const router = express.Router()
-
-// 3 600 000 is 1 hour
-const REFRESH = 3600000
-
-setInterval(() => {
-    updateNews()
-    console.log('ran') 
-}, REFRESH)
-
-function getWords(article) {
-    let words = []
-    retext().use(retextkeywords).process(
-        `${article.title} ${article.description}`, (err, file) => {
-            file.data.keywords.forEach((keyword) => {
-                // sanitize it
-                let word = sanitize(nlcstToString(keyword.matches[0].node))
-                words.push(nlcstToString(keyword.matches[0].node))
-            })
-        })
-    //console.log(article)
-    return words
-}
-
-function getPhrases(article) {
-    let words = []
-    retext().use(retextkeywords).process(
-        `${article.title} ${article.description}`, (err, file) => {
-            file.data.keyphrases.forEach((phrase) => {
-                words.push(sanitize(phrase.matches[0].nodes.map(nlcstToString).join('')))
-            })
-        })
-    //console.log(article)
-    return words
-}
-
-function sanitize(string) {
-    // formats the string to lowercase, and removes all punctuation
-    string = string.toLowerCase()
-    return string.replace(/[^0-9a-z ]/g, '')
-}
-
-function dropArticles() {
-    Article.remove({}, () => {})
-    Keyword.remove({}, () => {})
-}
-
-function updateNews() {
-    // hourly
-    const url = `${process.env.NEWS_URI}`
-    request(url)
-        .then((res) => {
-            const headlines = JSON.parse(res).articles
-            let articles = headlines.map((old) => {
-                // save art
-                Article.findOne({url: old.url})
-                    .then((article) => {
-                        // already in database, do nothing
-                        if (!article) throw new Error('Article not found')
-                    })
-                    .catch((err) => {
-                        // not in db
-                        let article = new Article(old)
-                        article['keywords'] = getPhrases(article)
-                        article['votes'] = 0
-                        console.log(article)
-                        article.save()
-                    })
-            })
-        })
-        .catch((err) => {
-            console.log(err)
-        })
-}
-
-function getAmount(vote) {
-    return Math.abs(vote) / vote || 0
-}
 
 router.post('/vote', (req, res) => {
     let amount = getAmount(req.body.vote)
@@ -211,6 +130,10 @@ function getNextArticles(findParams, sortParams) {
             .then(articles => resolve(articles))
             .catch(err => reject(err))
     })
+}
+
+function getAmount(vote) {
+    return Math.abs(vote) / vote || 0
 }
 
 module.exports = router
